started using fanstatic & extending templates properly
