+++ ckanext/dataproxy/templates/package/new_resource.html
@@ -1,26 +1,10 @@
-{% extends "package/base_form_page.html" %}
-
-{% set logged_in = true if c.userobj else false %}
-
-{% block subtitle %}{{ _('Add data to the dataset') }}{% endblock %}
-
-{% block breadcrumb_content_selected %}{% endblock %}
-{% block breadcrumb_content %}
+{% ckan_extends %}
+{% block styles %}
   {{ super() }}
-  {% if pkg %}
-    <li class="active"><a href="#">{{ _('Add New Resource') }}</a></li>
-  {% endif %}
+  {% resource 'dataproxy/dataproxy.css' %}
 {% endblock %}
-
-{% block form %}{% snippet 'package/snippets/resource_form.html', data=data, errors=errors, error_summary=error_summary, include_metadata=false, pkg_name=pkg_name, stage=stage, allow_upload=g.ofs_impl and logged_in %}{% endblock %}
-
-{% block secondary_content %}
-  {% snippet 'package/snippets/resource_help.html' %}
-{% endblock %}
-
 {% block scripts %}
   {{ super() }}
   {% resource 'vendor/fileupload' %}
-  <script type="text/javascript" src="{{ h.url_for_static('/dataproxy.js') }}" ></script>
-  <link rel="stylesheet" type="text/css" href="{{ h.url_for_static('/dataproxy.css') }}">
+  {% resource 'dataproxy/dataproxy.js' %}
 {% endblock %}
