+++ ckanext/dataproxy/templates/package/snippets/resource_edit_form.html
@@ -1,13 +1,2 @@
-{% set stage = stage or 1 %}
 {% set password_placeholder = _('Password hidden, enter new password only if you wish to change it.') %}
-{% set include_metadata = include_metadata or false %}
-
-{% extends 'package/snippets/resource_form.html' %}
-
-{% block stages %}{% endblock %}
-
-{% block previous_button %}{% endblock %}
-
-{% block again_button %}{% endblock %}
-
-{% block save_button_text %}{{ _('Update Resource') }}{% endblock %}
+{% ckan_extends %}
